After a day of working with Spark
Still not there, but I am able to add data to Postgres.  I'm not satified
with the way that the data is added to MySQL.  May want to have an intermediary
process that both creates the entities uniquely, and also the relationships to
the original stories that have been trimmed and stored to Postgres.